+++ lib/src/core/animations/animations.dart
@@ -1,7 +1,8 @@
-export 'package:flutter_easy_dialogs/src/core/animations/widgets/fullscreen_blur.dart';
+export 'package:flutter_easy_dialogs/src/core/animations/widgets/easy_full_screen_blur.dart';
 
 export 'easy_animation.dart';
 export 'easy_animation_settings.dart';
 export 'factory/i_easy_animation_factory.dart';
 export 'pre_built/pre_built.dart';
+export 'transitions/transitions.dart';
 export 'types/types.dart';